change port from int to uint16 (#102)

+++ cluster/config/slave_config.go
@@ -8,7 +8,7 @@ import (
 
 type SlaveConfig struct {
 	IP            string             `json:"HOST"` // DEFAULT_HOST
-	Port          int                `json:"PORT"` // 38392
+	Port          uint16             `json:"PORT"` // 38392
 	ID            string             `json:"ID"`
 	ChainMaskList []*types.ChainMask `json:"-"`
 }
